Improves lines editor file structure
